Refactor FMBLogo.js
