+++ src/components2/Logo/index.ts
@@ -1,0 +1,3 @@
+import Logo from './Logo';
+
+export default Logo;
